Blocking and related features

+++ server/controllers/event.controller.js
@@ -1,5 +1,30 @@
 const BaseEvent = require("../models/event");
 
+// Common function to authenticate the user
+async function authenticateUser(req) {
+    let authenticatedUser = null;
+    let authenticated = false;
+    const req_cookies = req.cookies;
+
+    if (req_cookies) {
+        const user_id = req_cookies.user_id;
+        if (user_id) {
+            authenticatedUser = await User.findById(user_id);
+            if (authenticatedUser) {
+                const auth_token = req_cookies.auth_token;
+                if (auth_token) {
+                    const userCredentials = authenticatedUser.userCredentials;
+                    if (userCredentials.matchAuthToken(auth_token)) {
+                        authenticated = true;
+                    }
+                }
+            }
+        }
+    }
+
+    return authenticated ? authenticatedUser : null;
+}
+
 // Retrieve an event by ID
 exports.getEventById = async (req, res) => {
     try {
@@ -24,21 +49,43 @@ exports.getEventById = async (req, res) => {
 // Create a new event
 exports.createEvent = async (req, res) => {
     try {
-        const { eventType, ...eventData } = req.body;
+        const authenticatedUser = await authenticateUser(req);
+        if (!authenticatedUser) {
+            return res.status(400).send({
+                message: "Not logged in!"
+            });
+        }
 
+        const { eventType, postId, ...eventData } = req.body;
+
+        // Check for event type
         if (!eventType) {
             return res.status(400).send({
                 message: "Event type is required."
             });
         }
 
+        // Check for post existence and event attachment
+        const post = await Post.findById(postId);
+        if (!post) {
+            return res.status(404).send({
+                message: "Post not found."
+            });
+        }
+
+        if (post.is_event) {
+            return res.status(400).send({
+                message: "An event is already attached to this post."
+            });
+        }
+
         let event;
         if (eventType === 'NormalEvent') {
-            event = new BaseEvent.discriminators.NormalEvent(eventData);
+            event = new BaseEvent.discriminators.NormalEvent({ ...eventData, post: postId });
         } else if (eventType === 'MusicReleaseEvent') {
-            event = new BaseEvent.discriminators.MusicReleaseEvent(eventData);
+            event = new BaseEvent.discriminators.MusicReleaseEvent({ ...eventData, post: postId });
         } else if (eventType === 'TicketedEvent') {
-            event = new BaseEvent.discriminators.TicketedEvent(eventData);
+            event = new BaseEvent.discriminators.TicketedEvent({ ...eventData, post: postId });
         } else {
             return res.status(400).send({
                 message: "Invalid event type."
@@ -46,6 +93,10 @@ exports.createEvent = async (req, res) => {
         }
 
         await event.save();
+        post.is_event = true;
+        post.eventId = event._id;
+        await post.save();
+
         res.status(201).send(event);
     } catch (err) {
         res.status(500).send({
@@ -58,15 +109,32 @@ exports.createEvent = async (req, res) => {
 // Edit an existing event
 exports.editEvent = async (req, res) => {
     try {
+        const authenticatedUser = await authenticateUser(req);
+        if (!authenticatedUser) {
+            return res.status(400).send({
+                message: "Not logged in!"
+            });
+        }
+
         const eventId = req.params.id;
         const updates = req.body;
 
-        const event = await BaseEvent.findByIdAndUpdate(eventId, updates, { new: true });
+        const event = await BaseEvent.findById(eventId);
         if (!event) {
             return res.status(404).send({
                 message: "Event not found."
             });
         }
+
+        const post = await Post.findById(event.post);
+        if (post.user.toString() !== authenticatedUser._id.toString()) {
+            return res.status(403).send({
+                message: "You are not allowed to edit someone else's event."
+            });
+        }
+
+        Object.assign(event, updates);
+        await event.save();
 
         res.status(200).send(event);
     } catch (err) {
@@ -80,14 +148,33 @@ exports.editEvent = async (req, res) => {
 // Delete an event
 exports.deleteEvent = async (req, res) => {
     try {
-        const eventId = req.params.id;
+        const authenticatedUser = await authenticateUser(req);
+        if (!authenticatedUser) {
+            return res.status(400).send({
+                message: "Not logged in!"
+            });
+        }
 
-        const event = await BaseEvent.findByIdAndDelete(eventId);
+        const eventId = req.params.id;
+        const event = await BaseEvent.findById(eventId);
+
         if (!event) {
             return res.status(404).send({
                 message: "Event not found."
             });
         }
+
+        const post = await Post.findById(event.post);
+        if (post.user.toString() !== authenticatedUser._id.toString()) {
+            return res.status(403).send({
+                message: "You are not allowed to delete someone else's event."
+            });
+        }
+
+        post.is_event = false;
+        post.eventId = null;
+        await post.save();
+        await BaseEvent.findByIdAndDelete(eventId);
 
         res.status(200).send({
             message: "Event successfully deleted."
